feat(solver/optimizer): add new file fsdp_optimizer.py
