+++ internlm/solver/optimizer/__init__.py
@@ -1,10 +1,7 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
 
-from .hybrid_zero_optim import (
-    FSDPadaptOptimizer,
-    HybridZeroOptimizer,
-    reload_zero_fp32_buff,
-)
+from .fsdp_optimizer import FSDPadaptOptimizer
+from .hybrid_zero_optim import HybridZeroOptimizer, reload_zero_fp32_buff
 
-__all__ = ["HybridZeroOptimizer", "FSDPadaptOptimizer", "reload_zero_fp32_buff"]
+__all__ = ["FSDPadaptOptimizer", "HybridZeroOptimizer", "reload_zero_fp32_buff"]
